Sum of int

+++ Calculator.java
@@ -1,6 +1,6 @@
 public class Calculator {
 
-public void sum(double a, double b){
+public void sum(int a, int b){
 	int sum = a + b;
 
 	System.out.println(sum);
